Moved sources into `src/printf/`, and installing the header under `printf/printf.h`, to allow inclusing using `<printf/printf.h>`, thus avoiding conflicts with glibc's `printf.h`.
